requirements.txt
pandas
numpy
scikit-learn
matplotlib
pillow
tkcalendar
scipy
google-generativeai
